Body, motion, and transform lists

RigidBodyTools.BodyList — Type

BodyList([b1,b2,...])

Create a list of bodies

RigidBodyTools.getrange — Function

getrange(bl::BodyList,i::Int) -> Range

Return the subrange of indices in the global set of surface point data corresponding to body i in a BodyList bl.

getrange(bl::BodyList,ml::MotionList,i::Int) -> Range

Return the subrange of indices in the global vector of motion state data corresponding to body i in a BodyList bl with corresponding motion list ml.

Base.collect — Method

collect(bl::bodylist) -> Vector{Float64}, Vector{Float64}

Collect the inertial-space coordinates of all of the Lagrange points comprising the bodies in body list bl and return each assembled set of coordinates as a vector.

Base.sum — Method

sum(f::AbstractVector,bl::BodyList,i::Int) -> Real

Compute a sum of the elements of vector f corresponding to body i in body list bl.

Base.view — Method

view(f::AbstractVector,bl::BodyList,i::Int) -> SubArray

Provide a view of the range of values in vector f corresponding to the Lagrange points of the body with index i in a BodyList bl.

RigidBodyTools.MotionList — Type

MotionList([m1,m2,...])

Create a list of motions

RigidBodyTools.RigidTransformList — Type

RigidTransformList([t1,t2,...])

Create a list of rigid transforms

Base.vec — Method

vec(tl::RigidTransformList) -> Vector{Float64}

Returns a concatenation of length-3 vectors of the form [x,y,α] corresponding to the translation and rotation specified by the given by the list of transforms tl.