Motions

Motion types

RigidBodyTools.BasicDirectMotion — Type

BasicDirectMotion(u::Vector{Float64},v::Vector{Float64})

Create an instance of basic directly-specified (constant) velocity, to be associated with a body whose length is the same as u and v.

RigidBodyTools.RigidBodyMotion — Type

RigidBodyMotion <: AbstractMotion

A type to store the body's current kinematics

Fields

c: current centroid position (relative to initial position)

ċ: current centroid velocity

c̈: current centroid acceleration

α: current angle (relative to initial angle)

α̇: current angular velocity

α̈: current angular acceleration

kin: a Kinematics structure

The first six fields are meant as a cache of the current kinematics while the kin field can be used to find the plate kinematics at any time.

RigidBodyTools.RigidBodyMotion — Method

RigidBodyMotion(kin::Kinematics)

Create an instance of rigid-body motion with kinematics kin.

RigidBodyTools.RigidBodyMotion — Method

RigidBodyMotion(ċ::ComplexF64,α̇)

Create an instance of constant rigid-body motion with velocity ċ and angular velocity α̇

RigidBodyTools.RigidAndDirectMotion — Type

RigidAndDirectMotion(rig::RigidBodyMotion,def::AbstractDirectlySpecifiedMotion)

Create an instance of basic superposition of a rigid-body motion and directly-specified deformation velocity in body coordinates.

RigidBodyTools.RigidAndDirectMotion — Method

RigidAndDirectMotion(kin::Kinematics,def::AbstractDirectlySpecifiedMotion)

Create an instance of basic superposition of a rigid-body motion with kinematics kin, and directly-specified deformation velocity in body coordinates.

RigidBodyTools.RigidAndDirectMotion — Method

RigidAndDirectMotion(kin::Kinematics,ũ::Vector{Float64},ṽ::Vector{Float64})

Create an instance of basic superposition of a rigid-body motion and directly-specified (constant) deformation velocity in body coordinates, to be associated with a body whose length is the same as ũ and ṽ.

RigidBodyTools.RigidAndDirectMotion — Method

RigidAndDirectMotion(ċ,α̇,ũ::Vector{Float64},ṽ::Vector{Float64})

Specify constant translational ċ and angular α̇ velocity and directly-specified (constant) deformation velocity in body coordinates, to be associated with a body whose length is the same as ũ and ṽ.

Surface velocity functions

RigidBodyTools.surface_velocity! — Function

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
             body::Body,motion::AbstractMotion,t::Real)

Assign the components of body velocity u and v (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in body in body at time t, based on supplied motions in the motion for the body.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 x::AbstractVector{Float64},y::AbstractVector{Float64},
                 xc::Real,yc::Real,α::Real,
                 motion::RigidBodyMotion,t::Real)

Assign the components of rigid body velocity u and v (in inertial coordinate system) at surface positions described by coordinates x, y (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 b::Body,motion::BasicDirectMotion,t::Real)

Assign the components of velocity u and v (in inertial coordinate system) at surface positions described by points in body b (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 b::Body,motion::RigidAndDirectMotion,t::Real)

Assign the components of velocity u and v (in inertial coordinate system) at surface positions described by points in body b (also in inertial coordinate system) at time t, based on supplied motion motion for the body. This function calls the supplied function for the deformation part in motion.defmotion.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 bl::BodyList,ml::MotionList,t::Real)

Assign the components of velocity u and v (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in each body in bl at time t, based on supplied motions in the MotionList ml for each body.

RigidBodyTools.surface_velocity — Function

surface_velocity(x::AbstractVector{Float64},y::AbstractVector{Float64},
                xc::Real,yc::Real,α::Real,motion::RigidBodyMotion,t::Real)

Return the components of rigid body velocities (in inertial components) at surface positions described by coordinates x, y (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

surface_velocity(bl::BodyList,ml::MotionList,t::Real)

Return the components of rigid body velocity (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in each body in bl at time t, based on supplied motions in the MotionList ml for each body.

surface_velocity(body::Body,motion::AbstractMotion,t::Real)

Return the components of rigid body velocity (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in body in body at time t, based on supplied motions in motion for the body.

Motion functions

RigidBodyTools.motion_state — Function

motion_state(b::Body,m::RigidBodyMotion)

Return the current state vector of body b associated with rigid body motion m. It returns the current coordinates of the body centroid and the angle of the body.

motion_state(b::Body,m::AbstractDirectlySpecifiedMotion)

Return the current state vector of body b associated with direct motion m. It returns the concatenated coordinates of the body surface (in the body-fixed coordinate system).

motion_state(b::Body,m::RigidAndDirectMotion)

Return the current state vector of body b associated with rigid+direct motion m. It returns the concatenated coordinates of the rigid-body mode and the body surface (in the body coordinate system).

motion_state(bl::BodyList,ml::MotionList)

Return the current state vector of body list bl associated with motion list ml. It returns the aggregated state vectors of each body.

RigidBodyTools.update_body! — Function

update_body!(b::Body,x::AbstractVector,m::RigidBodyMotion)

Update body b with the rigid-body motion state vector x. The information in m is used for checking lengths only.

update_body!(b::Body,x::AbstractVector,m::RigidAndDirectMotion)

Update body b with the motion state vector x, interpreted as coordinates in the body coordinate system. The information in m is used for parsing only.

update_body!(b::Body,x::AbstractVector,m::RigidAndDirectMotion)

Update body b with the motion state vector x. The part of the motion state associated with surface deformation is interpreted as expressed in body coordinates. The information in m is used for parsing only.

update_body!(bl::BodyList,x::AbstractVector,ml::MotionList)

Update the bodies in list bl with the given motion state vector x. The argument ml simply provides the information needed to parse the vector into each body.

RigidBodyTools.motion_velocity — Function

motion_velocity(b::Body,m::RigidBodyMotion,t::Real)

Return the velocity components (as a vector) of a RigidBodyMotion at the given time t.

motion_velocity(b::Body,m::AbstractDirectlySpecifiedMotion,t::Real)

Return the velocity components (as a vector) of a AbstractDirectlySpecifiedMotion at the given time t.

motion_velocity(b::Body,m::RigidAndDirectMotion,t::Real)

Return the velocity components (as a vector) of a RigidAndDirectMotion at the given time t.

motion_velocity(bl::BodyList,ml::MotionList,t::Real) -> Vector

Return the aggregated velocity components (as a vector) of a MotionList at the given time t.

Rigid body kinematics types

RigidBodyTools.Kinematics — Type

An abstract type for types that takes in time and returns (c, ċ, c̈, α, α̇, α̈).

RigidBodyTools.Oscillation — Type

Oscillation(Ux,Uy,α̇₀,ax,ay,Ω,Ax,Ay,ϕx,ϕy,α₀,Δα,ϕα) <: Kinematics

Set 2-d oscillatory kinematics. This general constructor sets up motion of a rotational axis, located at ax, ay (expressed relative to the body centroid, in a body-fixed coordinate system). The rotational axis motion is described by

$x(t) = U_x t + A_x\sin(\Omega t - \phi_x), \quad y(t) = U_y t + A_y\sin(\Omega t - \phi_y), \quad \alpha(t) = \alpha_0 + \dot{\alpha}_0 t + \Delta\alpha \sin(\Omega t - \phi_{\alpha})$

RigidBodyTools.OscillationX — Function

OscillationX(Ux,Ω,Ax,ϕx)

Set oscillatory kinematics in the $x$ direction, of the form

$x(t) = U_x t + A_x \sin(\Omega t - \phi_x),$

RigidBodyTools.OscillationY — Function

OscillationY(Uy,Ω,Ay,ϕy)

Set oscillatory kinematics in the $y$ direction, of the form

$y(t) = U_y t + A_y \sin(\Omega t - \phi_y)$

RigidBodyTools.OscillationXY — Function

OscillationXY(Ux,Uy,Ω,Ax,ϕx,Ay,ϕy)

Set oscillatory kinematics in the $x$ and $y$ directions, of the form

$x(t) = U_x t + A_x \sin(\Omega t - \phi_x), \quad y(t) = U_y t + A_y \sin(\Omega t - \phi_y)$

RigidBodyTools.PitchHeave — Function

PitchHeave(U₀,a,Ω,α₀,Δα,ϕp,A,ϕh)

Create oscillatory pitching and heaving kinematics of a pitch axis at location $a$ (expressed relative to the centroid in the $\tilde{x}$ direction of the body-fixed coordinate system), of the form (in inertial coordinates)

$x(t) = U_0, \quad y(t) = A\sin(\Omega t - \phi_h), \quad \alpha(t) = \alpha_0 + \Delta\alpha \sin(\Omega t - \phi_p)$

RigidBodyTools.Pitchup — Type

Pitchup(U₀,a,K,α₀,t₀,Δα,ramp=EldredgeRamp(11.0)) <: Kinematics

Kinematics describing a pitch-ramp motion (horizontal translation with rotation) starting at time $t_0$ about an axis at a (expressed relative to the centroid, in the $\tilde{x}$ direction in the body-fixed coordinate system), with translational velocity U₀ in the inertial $x$ direction, initial angle $\alpha_0$, dimensionless angular velocity $K = \dot{\alpha}_0c/2U_0$, and angular change $\Delta\alpha$. The optional ramp argument is assumed to be given by the smooth ramp EldredgeRamp with a smoothness factor of 11 (larger values lead to sharper transitions on/off the ramp), but this can be replaced by another Eldredge ramp with a different value or a ColoniusRamp.

RigidBodyTools.RotationalOscillation — Function

RotationalOscillation(ax,ay,Ω,α₀,α̇₀,Δα,ϕα)

Set oscillatory rotational kinematics about an axis located at ax, ay (expressed relative to the body centroid, in a body-fixed coordinate system), of the form

$\alpha(t) = \alpha_0 + \dot{\alpha}_0 t + \Delta\alpha \sin(\Omega t - \phi_{\alpha})$

RotationalOscillation(Ω,Δα,ϕα)

Set oscillatory rotational kinematics about the centroid of the form

$\alpha(t) = \Delta\alpha \sin(\Omega t - \phi_{\alpha})$

RigidBodyTools.SwitchedKinematics — Type

SwitchedKinematics <: Kinematics

Modulates a given set of kinematics between simple on/off states. The velocity specified by the given kinematics is toggled on/off.

Fields

t_on

time at which the kinematics should be turned on

t_off

time at which the kinematics should be turned off

kin

kinematics to be followed in the on state

off