Surface functions and other utilities

RigidBodyTools.centraldiff — Function

centraldiff(body::Body) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the circular central differences of coordinates on body body (or on each body in list body).

centraldiff(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the centraldiff on each constituent body in bl.

Base.diff — Method

diff(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y differences of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body (if ref=false), or at the reference x̃ and ỹ coordinates (body-fixed space) if ref=true. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

Base.diff — Method

diff(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the diff on each constituent body in bl.

RigidBodyTools.dlength — Function

dlength(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example. For an OpenBody, this provides a vector that is one element shorter than the number of points, to ensure that sum(dlength(body)) is equal to the arclength of the body.

RigidBodyTools.dlengthmid — Function

dlengthmid(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces formed between the face midpoints on the perimeter of body body. The indexing of these midpoint faces is consistent with that of the regular vertex points adjacent to both midpoints. Midpoint face 2 corresponds to the face between midpoints 1 and 2, for example. For an OpenBody, the lengths for the first and last points are calculated to the adjoining midpoints, to ensure that sum(dlength(body)) is equal to the arclength of the body.

Base.length — Method

length(body::Body)

Return the number of points on the body perimeter

RigidBodyTools.midpoints — Method

midpoints(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y midpoints of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body (if ref=false), or at the reference x̃ and ỹ coordinates (body-fixed space) if ref=true. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

RigidBodyTools.midpoints — Method

midpoints(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the midpoints on each constituent body in bl.

RigidBodyTools.normal — Function

normal(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals in inertial components (if ref=false) or body- fixed components (if ref=true) of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example. For an OpenBody, this provides a vector that is one element shorter than the number of points.

RigidBodyTools.normalmid — Function

normalmid(body::Body/BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals in inertial components (if ref=false) or body- fixed components (if ref=true) of the faces formed between midpoints on the perimeter of body body (or each body in list body). For an OpenBody, the normals for the first and last points are calculated for the face adjoining with the adjacent midpoints.