Index

Shapes

RigidBodyTools.BasicBody — Type

BasicBody(x,y[,closuretype=ClosedBody]) <: Body

Construct a body by simply passing in the x and y coordinate vectors. The last point will be automatically connected to the first point. The coordinate vectors are assumed to be expressed in the body-fixed coordinate system. The optional closuretype specifies whether the body is closed (ClosedBody) or open (OpenBody). If closed, then the first and last points are assumed joined in operations that require neighbor points.

RigidBodyTools.Circle — Function

Circle(a,n) <: Body

Construct a circular body with radius a and with n points distributed on the body perimeter.

Circle(a,targetsize::Float64) <: Body

Construct a circular body with radius a with spacing between points set approximately to targetsize.

RigidBodyTools.Ellipse — Type

Ellipse(a,b,n) <: Body

Construct an elliptical body with semi-major axis a and semi-minor axis b, with n points distributed on the body perimeter.

RigidBodyTools.NACA4 — Type

NACA4(cam,pos,thick,np,[len=1.0]) <: Body{N}

Generates points in the shape of a NACA 4-digit airfoil of chord length 1. The relative camber is specified by cam, the position of maximum camber (as fraction of chord) by pos, and the relative thickness by thick. The parameter np specifies the number of points on the upper or lower surface. The optional parameter len specifies the chord length, which defaults to 1.0.

Example

julia> w = Bodies.NACA4(0.0,0.0,0.12);

RigidBodyTools.Plate — Type

Plate(length,thick,n,[λ=1.0]) <: Body

Construct a flat plate with length length and thickness thick, with n points distributed on the body perimeter.

The optional parameter λ distributes the points differently. Values between 0.0 and 1.0 are accepted.

The constructor Plate(length,n,[λ=1.0]) creates a plate of zero thickness.

Alternatively, either form can be specified with a target spacing in place of n.

RigidBodyTools.Rectangle — Type

Rectangle(a,b,na) <: Body

Construct a rectangular body with x̃ side half-length a and ỹ side half-length b, with na points distributed on the x̃ side (including both corners). The centroid of the rectangle is placed at the origin (so that the lower left corner is at (-a,-b)).

By default, points are not placed at the corners, but rather, are shifted by half a segment. This ensures that all normals are perpendicular to the sides. If, instead, the shifted=false flag is added, then points are placed at the corners, shifted clockwise relative to the default by half a segment, and the normal vectors are bisectors between the normals on the adjacent two sides.

RigidBodyTools.SplinedBody — Function

SplinedBody(X,Δx[,closuretype=ClosedBody]) -> BasicBody

Using control points in X (assumed to be N x 2, where N is the number of points), create a set of points that are uniformly spaced (with spacing Δx) on a curve that passes through the control points. A cubic parametric spline algorithm is used. If the optional parameter closuretype is set to OpenBody, then the end points are not joined together.

RigidBodyTools.Square — Function

Square(a,na) <: Body

Construct a square body with side half-length a and with na points distributed on each side (including both corners).

Body list functions

RigidBodyTools.BodyList — Type

BodyList([b1,b2,...])

Create a list of bodies

RigidBodyTools.getrange — Function

getrange(bl::BodyList,i::Int) -> Range

Return the subrange of indices in the global set of surface point data corresponding to body i in a BodyList bl.

Base.collect — Method

collect(bl::bodylist) -> Vector{Float64}, Vector{Float64}

Collect the inertial-space coordinates of all of the Lagrange points comprising the bodies in body list bl and return each assembled set of coordinates as a vector.

Base.sum — Method

sum(f::AbstractVector,bl::BodyList,i::Int) -> Real

Compute a sum of the elements of vector f corresponding to body i in body list bl.

Base.view — Method

view(f::AbstractVector,bl::BodyList,i::Int) -> SubArray

Provide a view of the range of values in vector f corresponding to the Lagrange points of the body with index i in a BodyList bl.

Rigid transformations

RigidBodyTools.RigidTransform — Type

RigidTransform(x::Tuple{Float64,Float64},α::Float64)

Construct a rigid-body transform operator, with rotation by angle α and translation specified by x. The translation coordinates are specified in the target coordinate system.

The resulting transform can be used as an operator on pairs of coordinate vectors, x and y, or on bodies. For transformation of bodies, it only overwrites the x and y fields of the body, but leaves the x̃ and ỹ (body coordinates) intact.

The translation can be provided as either a tuple (x,y) or as a complex number.

Constructors

RigidTransform((x,y),α)

RigidTransform(u::Vector{Float64})

RigidTransform(u::NTuple{3,Float64})

RigidTransform.(u) where u is a collection of vectors or tuples.

Example

julia> body = RigidTransform.Ellipse(0.5,0.1,100)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

julia> T = RigidTransform((1.0,1.0),π/4)
Rigid-body transform
  Translation: (1.0,1.0)
  Rotation angle (rad): 0.7853981633974483

julia> T(body)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (1.0,1.0)
   Current angle (rad): 0.7853981633974483

RigidBodyTools.RigidTransformList — Type

RigidTransformList([t1,t2,...])

Create a list of rigid transforms

Base.vec — Method

vec(T::RigidTransform) -> Vector{Float64}

Returns a length-3 vector of the form [x,y,α] corresponding to the translation and rotation specified by the given transform T.

Base.vec — Method

vec(tl::RigidTransformList) -> Vector{Float64}

Returns a concatenation of length-3 vectors of the form [x,y,α] corresponding to the translation and rotation specified by the given by the list of transforms tl.

Surface functions

RigidBodyTools.centraldiff — Function

centraldiff(body::Body) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the circular central differences of coordinates on body body (or on each body in list body).

centraldiff(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the centraldiff on each constituent body in bl.

Base.diff — Method

diff(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y differences of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body (if ref=false), or at the reference x̃ and ỹ coordinates (body-fixed space) if ref=true. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

Base.diff — Method

diff(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the diff on each constituent body in bl.

RigidBodyTools.dlength — Function

dlength(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example. For an OpenBody, this provides a vector that is one element shorter than the number of points, to ensure that sum(dlength(body)) is equal to the arclength of the body.

RigidBodyTools.dlengthmid — Function

dlengthmid(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces formed between the face midpoints on the perimeter of body body. The indexing of these midpoint faces is consistent with that of the regular vertex points adjacent to both midpoints. Midpoint face 2 corresponds to the face between midpoints 1 and 2, for example. For an OpenBody, the lengths for the first and last points are calculated to the adjoining midpoints, to ensure that sum(dlength(body)) is equal to the arclength of the body.

Base.length — Method

length(body::Body)

Return the number of points on the body perimeter

RigidBodyTools.midpoints — Function

midpoints(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y midpoints of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body (if ref=false), or at the reference x̃ and ỹ coordinates (body-fixed space) if ref=true. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

midpoints(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the midpoints on each constituent body in bl.

RigidBodyTools.normal — Function

normal(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals in inertial components (if ref=false) or body- fixed components (if ref=true) of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example. For an OpenBody, this provides a vector that is one element shorter than the number of points.

RigidBodyTools.normalmid — Function

normalmid(body::Body/BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals in inertial components (if ref=false) or body- fixed components (if ref=true) of the faces formed between midpoints on the perimeter of body body (or each body in list body). For an OpenBody, the normals for the first and last points are calculated for the face adjoining with the adjacent midpoints.

Motion functions

RigidBodyTools.BasicDirectMotion — Type

BasicDirectMotion(u::Vector{Float64},v::Vector{Float64})

Create an instance of basic directly-specified (constant) velocity, to be associated with a body whose length is the same as u and v.

RigidBodyTools.RigidBodyMotion — Type

RigidBodyMotion <: AbstractMotion

A type to store the body's current kinematics

Fields

c: current centroid position (relative to initial position)

ċ: current centroid velocity

c̈: current centroid acceleration

α: current angle (relative to initial angle)

α̇: current angular velocity

α̈: current angular acceleration

kin: a Kinematics structure

The first six fields are meant as a cache of the current kinematics while the kin field can be used to find the plate kinematics at any time.

RigidBodyTools.MotionList — Type

MotionList([m1,m2,...])

Create a list of motions

RigidBodyTools.motion_state — Function

motion_state(b::Body,m::RigidBodyMotion)

Return the current state vector of body b associated with rigid body motion m. It returns the current coordinates of the body centroid and the angle of the body.

motion_state(b::Body,m::DirectlySpecifiedMotion)

Return the current state vector of body b associated with direct motion m. It returns the concatenated coordinates of the body surface (in the inertial coordinate system).

motion_state(bl::BodyList,ml::MotionList)

Return the current state vector of body list bl associated with motion list ml. It returns the aggregated state vectors of each body.

RigidBodyTools.motion_velocity — Function

motion_velocity(b::Body,m::RigidBodyMotion,t::Real)

Return the velocity components (as a vector) of a RigidBodyMotion at the given time t.

motion_velocity(b::Body,m::DirectlySpecifiedMotion,t::Real)

Return the velocity components (as a vector) of a DirectlySpecifiedMotion at the given time t.

motion_velocity(bl::BodyList,ml::MotionList,t::Real) -> Vector

Return the aggregated velocity components (as a vector) of a MotionList at the given time t.

RigidBodyTools.surface_velocity! — Function

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
             body::Body,motion::AbstractMotion,t::Real)

Assign the components of body velocity u and v (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in body in body at time t, based on supplied motions in the motion for the body.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 x::AbstractVector{Float64},y::AbstractVector{Float64},
                 xc::Real,yc::Real,α::Real,
                 motion::RigidBodyMotion,t::Real)

Assign the components of rigid body velocity u and v (in inertial coordinate system) at surface positions described by coordinates x, y (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 b::Body,motion::BasicDirectMotion,t::Real)

Assign the components of velocity u and v (in inertial coordinate system) at surface positions described by points in body b (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

surface_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 bl::BodyList,ml::MotionList,t::Real)

Assign the components of velocity u and v (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in each body in bl at time t, based on supplied motions in the MotionList ml for each body.

RigidBodyTools.surface_velocity — Function

surface_velocity(x::AbstractVector{Float64},y::AbstractVector{Float64},
                xc::Real,yc::Real,α::Real,motion::RigidBodyMotion,t::Real)

Return the components of rigid body velocities (in inertial components) at surface positions described by coordinates x, y (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

surface_velocity(bl::BodyList,ml::MotionList,t::Real)

Return the components of rigid body velocity (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in each body in bl at time t, based on supplied motions in the MotionList ml for each body.

surface_velocity(body::Body,motion::AbstractMotion,t::Real)

Return the components of rigid body velocity (in inertial coordinate system) at surface positions described by coordinates inertial coordinates in body in body at time t, based on supplied motions in motion for the body.

Rigid body kinematics types

RigidBodyTools.Kinematics — Type

An abstract type for types that takes in time and returns (c, ċ, c̈, α, α̇, α̈).

RigidBodyTools.Oscillation — Type

Oscillation(Ux,Uy,α̇₀,ax,ay,Ω,Ax,Ay,ϕx,ϕy,α₀,Δα,ϕα) <: Kinematics

Set 2-d oscillatory kinematics. This general constructor sets up motion of a rotational axis, located at ax, ay (expressed relative to the body centroid, in a body-fixed coordinate system). The rotational axis motion is described by

$x(t) = U_x t + A_x\sin(\Omega t - \phi_x), \quad y(t) = U_y t + A_y\sin(\Omega t - \phi_y), \quad \alpha(t) = \alpha_0 + \dot{\alpha}_0 t + \Delta\alpha \sin(\Omega t - \phi_{\alpha})$

RigidBodyTools.OscillationX — Function

OscillationX(Ux,Ω,Ax,ϕx)

Set oscillatory kinematics in the $x$ direction, of the form

$x(t) = U_x t + A_x \sin(\Omega t - \phi_x),$

RigidBodyTools.OscillationY — Function

OscillationY(Uy,Ω,Ay,ϕy)

Set oscillatory kinematics in the $y$ direction, of the form

$y(t) = U_y t + A_y \sin(\Omega t - \phi_y)$

RigidBodyTools.OscillationXY — Function

OscillationXY(Ux,Uy,Ω,Ax,ϕx,Ay,ϕy)

Set oscillatory kinematics in the $x$ and $y$ directions, of the form

$x(t) = U_x t + A_x \sin(\Omega t - \phi_x), \quad y(t) = U_y t + A_y \sin(\Omega t - \phi_y)$

RigidBodyTools.PitchHeave — Function

PitchHeave(U₀,a,Ω,α₀,Δα,ϕp,A,ϕh)

Create oscillatory pitching and heaving kinematics of a pitch axis at location $a$ (expressed relative to the centroid in the $\tilde{x}$ direction of the body-fixed coordinate system), of the form (in inertial coordinates)

$x(t) = U_0, \quad y(t) = A\sin(\Omega t - \phi_h), \quad \alpha(t) = \alpha_0 + \Delta\alpha \sin(\Omega t - \phi_p)$

RigidBodyTools.Pitchup — Type

Pitchup(U₀,a,K,α₀,t₀,Δα,ramp=EldredgeRamp(11.0)) <: Kinematics

Kinematics describing a pitch-ramp motion (horizontal translation with rotation) starting at time $t_0$ about an axis at a (expressed relative to the centroid, in the $\tilde{x}$ direction in the body-fixed coordinate system), with translational velocity U₀ in the inertial $x$ direction, initial angle $\alpha_0$, dimensionless angular velocity $K = \dot{\alpha}_0c/2U_0$, and angular change $\Delta\alpha$. The optional ramp argument is assumed to be given by the smooth ramp EldredgeRamp with a smoothness factor of 11 (larger values lead to sharper transitions on/off the ramp), but this can be replaced by another Eldredge ramp with a different value or a ColoniusRamp.

RigidBodyTools.RotationalOscillation — Function

RotationalOscillation(ax,ay,Ω,α₀,α̇₀,Δα,ϕα)

Set oscillatory rotational kinematics about an axis located at ax, ay (expressed relative to the body centroid, in a body-fixed coordinate system), of the form

$\alpha(t) = \alpha_0 + \dot{\alpha}_0 t + \Delta\alpha \sin(\Omega t - \phi_{\alpha})$

RotationalOscillation(Ω,Δα,ϕα)

Set oscillatory rotational kinematics about the centroid of the form

$\alpha(t) = \Delta\alpha \sin(\Omega t - \phi_{\alpha})$

RigidBodyTools.SwitchedKinematics — Type

SwitchedKinematics <: Kinematics

Modulates a given set of kinematics between simple on/off states. The velocity specified by the given kinematics is toggled on/off.

Fields

t_on

time at which the kinematics should be turned on

t_off

time at which the kinematics should be turned off

kin

kinematics to be followed in the on state

off