Index

RigidBodyTools.BasicBody — Type

BasicBody(x,y) <: Body

Construct a body by simply passing in the x and y coordinate vectors. The last point will be automatically connected to the first point. The coordinate vectors are assumed to be expressed in the body-fixed coordinate system.

RigidBodyTools.Ellipse — Type

Ellipse(a,b,n) <: Body

Construct an elliptical body with semi-major axis a and semi-minor axis b, with n points distributed on the body perimeter.

RigidBodyTools.Kinematics — Type

An abstract type for types that takes in time and returns (c, ċ, c̈, α, α̇, α̈).

RigidBodyTools.NACA4 — Type

NACA4(cam,pos,thick[;np=20][,Xc=(0.0,0.0)][,len=1.0]) <: Body{N}

Generates points in the shape of a NACA 4-digit airfoil of chord length 1. The relative camber is specified by cam, the position of maximum camber (as fraction of chord) by pos, and the relative thickness by thick.

The optional parameter np specifies the number of points on the upper or lower surface. The optional parameter Zc specifies the mean position of the vertices (which is set to the origin by default). The optional parameter len specifies the chord length.

Example

julia> w = Bodies.NACA4(0.0,0.0,0.12);

RigidBodyTools.Plate — Type

Plate(length,thick,n,[λ=1.0]) <: Body

Construct a flat plate with length length and thickness thick, with n points distributed on the body perimeter.

The optional parameter λ distributes the points differently. Values between 0.0 and 1.0 are accepted.

The constructor Plate(length,n,[λ=1.0]) creates a plate of zero thickness.

RigidBodyTools.Rectangle — Type

Rectangle(a,b,na) <: Body

Construct a rectangular body with x̃ side half-length a and ỹ side half-length b, with na points distributed on the x̃ side (including both corners). The centroid of the rectangle is placed at the origin (so that the lower left corner is at (-a,-b)).

RigidBodyTools.RigidBodyMotion — Type

RigidBodyMotion

A type to store the body's current kinematics

Fields

c: current centroid position (relative to initial position)

ċ: current centroid velocity

c̈: current centroid acceleration

α: current angle (relative to initial angle)

α̇: current angular velocity

α̈: current angular acceleration

kin: a Kinematics structure

The first six fields are meant as a cache of the current kinematics while the kin field can be used to find the plate kinematics at any time.

RigidBodyTools.RigidTransform — Type

RigidTransform(x::Tuple{Float64,Float64},α::Float64)

Construct a rigid-body transform operator, with rotation by angle α and translation specified by x. The translation coordinates are specified in the target coordinate system.

The resulting transform can be used as an operator on pairs of coordinate vectors, x and y, or on bodies. For transformation of bodies, it only overwrites the x and y fields of the body, but leaves the x̃ and ỹ (body coordinates) intact.

The translation can be provided as either a tuple (x,y) or as a complex number.

Constructors

RigidTransform((x,y),α)

RigidTransform(u::Vector{Float64})

RigidTransform(u::NTuple{3,Float64})

RigidTransform.(u) where u is a collection of vectors or tuples.

Example

julia> body = Bodies.Ellipse(0.5,0.1,100)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

julia> T = RigidTransform((1.0,1.0),π/4)
Rigid-body transform
  Translation: (1.0,1.0)
  Rotation angle (rad): 0.7853981633974483

julia> T(body)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (1.0,1.0)
   Current angle (rad): 0.7853981633974483

RigidBodyTools.Circle — Method

Circle(a,n) <: Body

Construct a circular body with radius a and with n points distributed on the body perimeter.

RigidBodyTools.SplinedBody — Method

SplinedBody(X,Δx) -> BasicBody

Using control points in X (assumed to be N x 2, where N is the number of points), create a set of points that are uniformly spaced (with spacing Δx) on a curve that passes through the control points. A cubic parametric spline algorithm is used.

RigidBodyTools.Square — Method

Square(a,na) <: Body

Construct a square body with side half-length a and with na points distributed on each side (including both corners).

RigidBodyTools.assign_velocity! — Method

assign_velocity!(u::AbstractVector{Float64},v::AbstractVector{Float64},
                 x::AbstractVector{Float64},y::AbstractVector{Float64},
                 xc::Real,yc::Real,α::Real,
                 motion,t::Real)

Assign the components of rigid body velocity u and v (in inertial coordinate system) at positions described by coordinates x, y (also in inertial coordinate system) at time t, based on supplied motion motion for the body.

RigidBodyTools.centraldiff — Method

centraldiff(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the circular central differences of coordinates on body body (or on each body in list body).

RigidBodyTools.d_dt — Method

d_dt(p::Profile)

Take the time derivative of p and return it as a new profile.

Example

julia> s = Plates.RigidBodyMotions.Sinusoid(π)
Sinusoid (ω = 3.14)

julia> s.([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 0.0
 1.0
 0.707107

julia> c = Plates.RigidBodyMotions.d_dt(s)
d/dt (Sinusoid (ω = 3.14))

julia> c.([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
  3.14159
  1.92367e-16
 -2.22144

RigidBodyTools.dlength — Method

dlength(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

RigidBodyTools.dlengthmid — Method

dlengthmid(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces formed between the face midpoints on the perimeter of body body. The indexing of these midpoint faces is consistent with that of the regular vertex points adjacent to both midpoints. Midpoint face 2 corresponds to the face between midpoints 1 and 2, for example.

RigidBodyTools.getrange — Method

getrange(bl::BodyList,i::Int) -> Range

Return the range of indices in the global set of Lagrange point data corresponding to body i in body list bl.

RigidBodyTools.midpoints — Method

midpoints(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y midpoints of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

RigidBodyTools.normal — Method

normal(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals (in inertial components) of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

RigidBodyTools.normalmid — Method

normalmid(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals (in inertial components) of the faces formed between midpoints on the perimeter of body body (or each body in list body).

RigidBodyTools.ConstantProfile — Type

ConstantProfile(c::Number)

Create a profile consisting of a constant c.

Example

julia> p = RigidBodyMotions.ConstantProfile(1.0)
Constant (2.3)

RigidBodyTools.PitchHeave — Type

PitchHeave <: Kinematics

Kinematics describing an oscillatory pitching and heaving (i.e. plunging) motion

Constructors

Fields

U₀

Freestream velocity

a

Axis of pitch rotation, relative to the plate centroid

K

Reduced frequency $K = \frac{\Omega c}{2U_0}$

ϕp

Phase of pitch (in radians)

ϕh

Phase of heave (in radians)

α₀

Mean angle of attack

Δα

Amplitude of pitching

A

Amplitude of translational heaving

Y

Ẏ

Ÿ

α

α̇

α̈

RigidBodyTools.Pitchup — Type

Pitchup <: Kinematics

Kinematics describing a pitchup motion (horizontal translation with rotation)

Constructors

Fields

U₀

Freestream velocity

a

Axis of rotation, relative to the plate centroid

K

Non-dimensional pitch rate $K = \dot{\alpha}_0\frac{c}{2U_0}$

α₀

Initial angle of attack

t₀

Nominal start of pitch up

Δα

Total pitching angle

α

α̇

α̈

RigidBodyTools.Profile — Type

An abstract type for real-valued functions of time.

RigidBodyTools.SwitchedKinematics — Type

SwitchedKinematics <: Kinematics

Modulates a given set of kinematics between simple on/off states. The velocity specified by the given kinematics is toggled on/off.

Fields

t_on

time at which the kinematics should be turned on

t_off

time at which the kinematics should be turned off

kin

kinematics to be followed in the on state

off

Base.:* — Method

s::Number * p::Profile

Returns a scaled profile with (s*p)(t) = s*p(t)

Example

julia> s = Plates.RigidBodyMotions.Sinusoid(π)
Sinusoid (ω = 3.14)

julia> 2s
2 × (Sinusoid (ω = 3.14))

julia> (2s).([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 0.0
 2.0
 1.41421

Base.:+ — Method

p₁::Profile + p₂::Profile

Add the profiles so that (p₁ + p₂)(t) = p₁(t) + p₂(t).

Examples

julia> ramp₁ = Plates.RigidBodyMotions.EldredgeRamp(5)
logcosh ramp (aₛ = 5.0)

julia> ramp₂ = Plates.RigidBodyMotions.ColoniusRamp(5)
power series ramp (n = 5.0)

julia> ramp₁ + ramp₂
AddedProfiles:
  logcosh ramp (aₛ = 5.0)
  power series ramp (n = 5.0)

julia> ramp₁ + (ramp₂ + ramp₁) == ramp₁ + ramp₂ + ramp₁
true

Base.:- — Method

-(p₁::Profile, p₂::Profile)

julia> s = Plates.RigidBodyMotions.Sinusoid(π)
Sinusoid (ω = 3.14)

julia> 2s
2 × (Sinusoid (ω = 3.14))

julia> (2s).([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 0.0
 2.0
 1.41421

julia> s = Plates.RigidBodyMotions.Sinusoid(π);

julia> s.([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 0.0
 1.0
 0.707107

julia> (-s).([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 -0.0
 -1.0
 -0.707107

julia> (s - s).([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

Base.:>> — Method

p::Profile >> Δt::Number

Shift the profile in time so that (p >> Δt)(t) = p(t - Δt)

Example

julia> s = Plates.RigidBodyMotions.Sinusoid(π);

julia> s >> 0.5
Sinusoid (ω = 3.14) >> 0.5

julia> (s >> 0.5).([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
 -1.0
  0.0
  0.707107

julia> (s << 0.5).([0.0, 0.5, 0.75])
3-element Array{Float64,1}:
  1.0
  1.22465e-16
 -0.707107

Base.collect — Method

collect(bl::bodylist) -> Vector{Float64}, Vector{Float64}

Collect the inertial-space coordinates of all of the Lagrange points comprising the bodies in body list bl and return each assembled set of coordinates as a vector.

Base.diff — Method

diff(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y differences of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

Base.length — Method

length(body::Body)

Return the number of points on the body perimeter

Base.sum — Method

sum(f::AbstractVector,bl::BodyList,i::Int) -> Real

Compute a sum of the elements of vector f corresponding to body i in body list bl.

Base.vec — Method

vec(T::RigidTransform) -> Vector{Float64}

Returns a length-3 vector of the form [x,y,α] corresponding to the translation and rotation specified by the given transform T.

Base.vec — Method

vec(tl::Vector{RigidTransform}) -> NTuple{N,Vector{Float64}}

Returns a tuple of length-3 vectors of the form [x,y,α] corresponding to the translation and rotation specified by the given by the list of transforms tl.

Base.view — Method

view(f::AbstractVector,bl::BodyList,i::Int) -> SubArray

Provide a view of the range of values in vector f corresponding to the Lagrange points of the body with index i in body list bl.