Bodies

Shapes

RigidBodyTools.BasicBody — Type

BasicBody(x,y[,closuretype=ClosedBody]) <: Body

Construct a body by simply passing in the x and y coordinate vectors. The last point will be automatically connected to the first point. The coordinate vectors are assumed to be expressed in the body-fixed coordinate system. The optional closuretype specifies whether the body is closed (ClosedBody) or open (OpenBody). If closed, then the first and last points are assumed joined in operations that require neighbor points.

RigidBodyTools.Circle — Function

Circle(a,n) <: Body

Construct a circular body with radius a and with n points distributed on the body perimeter.

Circle(a,targetsize::Float64) <: Body

Construct a circular body with radius a with spacing between points set approximately to targetsize.

RigidBodyTools.Ellipse — Type

Ellipse(a,b,n) <: Body

Construct an elliptical body with semi-major axis a and semi-minor axis b, with n points distributed on the body perimeter.

RigidBodyTools.NACA4 — Type

NACA4(cam,pos,thick,np,[len=1.0]) <: Body{N}

Generates points in the shape of a NACA 4-digit airfoil of chord length 1. The relative camber is specified by cam, the position of maximum camber (as fraction of chord) by pos, and the relative thickness by thick. The parameter np specifies the number of points on the upper or lower surface. The optional parameter len specifies the chord length, which defaults to 1.0.

Example

julia> w = Bodies.NACA4(0.0,0.0,0.12);

RigidBodyTools.Plate — Type

Plate(length,thick,n,[λ=1.0]) <: Body

Construct a flat plate with length length and thickness thick, with n points distributed on the body perimeter.

The optional parameter λ distributes the points differently. Values between 0.0 and 1.0 are accepted.

The constructor Plate(length,n,[λ=1.0]) creates a plate of zero thickness.

Alternatively, either form can be specified with a target spacing in place of n.

RigidBodyTools.Rectangle — Type

Rectangle(a,b,na) <: Body

Construct a rectangular body with x̃ side half-length a and ỹ side half-length b, with na points distributed on the x̃ side (including both corners). The centroid of the rectangle is placed at the origin (so that the lower left corner is at (-a,-b)).

By default, points are not placed at the corners, but rather, are shifted by half a segment. This ensures that all normals are perpendicular to the sides. If, instead, the shifted=false flag is added, then points are placed at the corners, shifted clockwise relative to the default by half a segment, and the normal vectors are bisectors between the normals on the adjacent two sides.

RigidBodyTools.SplinedBody — Function

SplinedBody(X,Δx[,closuretype=ClosedBody]) -> BasicBody

Using control points in X (assumed to be N x 2, where N is the number of points), create a set of points that are uniformly spaced (with spacing Δx) on a curve that passes through the control points. A cubic parametric spline algorithm is used. If the optional parameter closuretype is set to OpenBody, then the end points are not joined together.

RigidBodyTools.Square — Function

Square(a,na) <: Body

Construct a square body with side half-length a and with na points distributed on each side (including both corners).

Rigid transformations of shapes

RigidBodyTools.RigidTransform — Type

RigidTransform(x::Tuple{Float64,Float64},α::Float64)

Construct a rigid-body transform operator, with rotation by angle α and translation specified by x. The translation coordinates are specified in the target coordinate system.

The resulting transform can be used as an operator on pairs of coordinate vectors, x and y, or on bodies. For transformation of bodies, it only overwrites the x and y fields of the body, but leaves the x̃ and ỹ (body coordinates) intact.

The translation can be provided as either a tuple (x,y) or as a complex number.

Constructors

RigidTransform((x,y),α)

RigidTransform(u::Vector{Float64})

RigidTransform(u::NTuple{3,Float64})

RigidTransform.(u) where u is a collection of vectors or tuples.

Example

julia> body = RigidTransform.Ellipse(0.5,0.1,100)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

julia> T = RigidTransform((1.0,1.0),π/4)
Rigid-body transform
  Translation: (1.0,1.0)
  Rotation angle (rad): 0.7853981633974483

julia> T(body)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (1.0,1.0)
   Current angle (rad): 0.7853981633974483

Base.vec — Method

vec(T::RigidTransform) -> Vector{Float64}

Returns a length-3 vector of the form [x,y,α] corresponding to the translation and rotation specified by the given transform T.