Creating and transforming bodies

The most basic functions of this package create an object of type Body. There are a variety of such functions, a few of which we will demonstrate here. Generally speaking, we are interesting in creating the object and placing it in a certain position and orientation. We do this in two steps: we create the basic shape, centered at the origin with a default orientation, and then we transform the shape to a desired location and orientation using a RigidTransform.

It is useful to stress that each body stores two types of points internally: the endpoints of the segments that comprise the body surface, and the midpoints of these segments. The midpoints are intended for use in downstream calculations, e.g. as forcing points in the calculations on immersed layers. The midpoints are simply the geometric averages of the endpoints, so endpoints are the ones that are transformed first, and midpoints are updated next.

using RigidBodyTools
using Plots

Creating a shape

Let's first create a shape. For any shape, we have to make a choice of the geometric dimensions (e.g, radius of the circle, side lengths of a rectangle), as well as the points that we use to discretely represent the surface. For this latter choice, there are two constructor types: we can specify the number of points (as an integer), or we can specify the nominal spacing between points (as a floating-point number).

The second approach is usually preferable when we use these tools for constructing immersed bodies. It is important to stress that the algorithms for placing points attempt to make the spacing as uniform as possible.

Let's create the most basic shape, a circle of radius 1. We will discretize it with 100 points first:

b = Circle(1.0,100)

Circular body with 100 points and radius 1.0
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

Now we will create the same body with a spacing of 0.02

b = Circle(1.0,0.02)

Circular body with 312 points and radius 1.0
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

This choice led to 312 points along the circumference. Quick math will tell you that the point spacing is probably not exactly 0.02. In fact, you can find out the actual spacing with dlengthmid. This function calculates the spacing associated with each point. (It does so by first calculating the spacing between midpoints between each point and its two adjacent points.)

dlengthmid(b)

312-element Vector{Float64}:
 0.020139095077867236
 0.020139094653432214
 0.020139095926408276
 0.020139093805405513
 0.02013909677357474
 0.02013909295944404
 0.020139097617989644
 0.02013909211691784
 0.020139098458281813
 0.020139091279197854
 ⋮
 0.02013909887637942
 0.02013909169745211
 0.020139098038654317
 0.020139092537748232
 0.02013909719612995
 0.020139093382162535
 0.02013909635016884
 0.020139094229325732
 0.0201390955021457

It is just a bit larger than 0.02.

A few other useful functions on the shape. To simply know the number of points,

length(b)

312

To find the outward normal vectors (based on the perpendicular to the line joining the adjacent midpoints):

nx, ny = normalmid(b)

([0.9999999999994525, 0.9997972500513199, 0.9991889560711672, 0.9981756172871893, 0.9967572242069976, 0.9949349397149282, 0.9927087488883388, 0.990080473523142, 0.9870500969577606, 0.983620092474754  …  0.9797908571175992, 0.9836197214184055, 0.9870504283171837, 0.9900801823981026, 0.9927089993067706, 0.9949347304092768, 0.9967573920603715, 0.9981754911583557, 0.9991890402708863, 0.9997972079172877], [1.0463938992680292e-6, 0.020136007295848538, 0.04026698480655045, 0.060377454843166735, 0.08046760833503637, 0.10052097161514166, 0.1205377114456275, 0.14050144464805614, 0.16041230032536535, 0.18025402541956284  …  -0.20002468924800312, -0.1802560502085253, -0.16041026138892794, -0.14050349612212365, -0.1205356490642922, -0.10052304325585998, -0.08046552909791128, -0.060379540001355994, -0.04026489541207215, -0.020138099234931442])

We can also plot the shape

plot(b)

Sometimes we don't want to fill in the shape (and maybe change the line color). In that case, we can use

plot(b,fill=:false,linecolor=:black)

Transforming the shape

Now, suppose we wish to place a shape at a different spot, with a different orientation. For this example, we will use a rectangle, since circles aren't much help for demonstrating orientation changes.

b = Rectangle(2.0,1.0,0.02)
plot(b)

Okay, now let's place it at (-1,-3), with an angle of π/3. We first create a RigidTransform.

T = RigidTransform((-1.0,-3.0),π/3)

Rigid-body transform
  Translation: (-1.0,-3.0)
  Rotation angle (rad): 1.0471975511965976

Now we apply this operator. The object T is function-like, and modifies the body in place:

T(b)

Closed polygon with 4 vertices and 600 points
   Current position: (-1.0,-3.0)
   Current angle (rad): 1.0471975511965976

Now b is transformed. Let's plot it:

plot(b)

There is an important thing to stress here. Each body keeps two sets of coordinates for the surface points. One that describe the shape in the original configuration, which we will refer to as the body coordinate system (or reference coordinates), and another set that describes the shape in the inertial coordinate system, which we, the viewers, are in. Only the second set of coordinates have been changed by the transform.

This also means that, when we apply another transform to the body, it is not a composite operation; it applies the transform to the reference shape, not the current shape.

Sometimes we need information about the normals in the reference system. For these, we can use normalmid with the flag ref=true:

nx, ny = normalmid(b,ref=true)

([0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  -1.0000000000000027, -1.0000000000000122, -0.9999999999999803, -1.0000000000000027, -0.9999999999999899, -1.000000000000001, -1.0000000000000027, -1.0000000000000027, -1.0000000000000122, -1.0000000000000027], [-1.0000000000000355, -0.999999999999997, -0.999999999999997, -0.999999999999997, -0.999999999999997, -1.0000000000000244, -0.9999999999999643, -1.00000000000003, -0.9999999999999643, -1.00000000000003  …  -0.0, -0.0, -0.0, -0.0, -0.0, -0.0, -0.0, -0.0, -0.0, -0.0])

Other shapes

Let's see some other shapes in action, like a square and an ellipse

b1, b2 = Square(1.0,0.02), Ellipse(0.6,0.1,0.02)
plot(plot(b1), plot(b2))

A NACA 4412 airfoil, with chord length 1, and 0.02 spacing between points, which we will place at 20 degrees angle of attack

b = NACA4(0.04,0.4,0.12,0.02)
T = RigidTransform((0.0,0.0),-20π/180)
T(b)
plot(b)

A flat plate with no thickness, at 45 degrees angle of attack

b = Plate(1.0,0.02)
T = RigidTransform((0.0,0.0),-45π/180)
T(b)
plot(b)

and a flat plate with a 5 percent thickness (and rounded ends)

b = ThickPlate(1.0,0.05,0.01)
plot(b)

There are also some generic tools for creating shapes. A BasicBody simply consists of points that describe the vertices. The interface for this is very simple.

x = [1.0, 1.2, 0.7, 0.6, 0.2, -0.1, 0.1, 0.4]
y = [0.1, 0.5, 0.8, 1.2, 0.8, 0.6, 0.2, 0.3]
b = BasicBody(x,y)

Basic pointwise-specified body with 8 points
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

plot(b)
scatter!(b,markersize=3,markercolor=:black)

However, this function does not insert any points along the sides between vertices. We have to do the work of specifying these points in the original call. For this reason, there are a few functions that are more directly useful. For example, we can create a polygon from these vertices, with a specified spacing between points distributed along the polygon sides

b = Polygon(x,y,0.02)

Closed polygon with 8 vertices and 188 points
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

plot(b)
scatter!(b,markersize=3,markercolor=:black)

Alternatively, we can interpret those original points as control points for splines, with a spacing between points along the splines provided:

b = SplinedBody(x,y,0.02)

Basic pointwise-specified body with 198 points
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

plot(b)
scatter!(b,markersize=3,markercolor=:black)

Body functions

RigidBodyTools.BasicBody — Type

BasicBody(x,y[,closuretype=ClosedBody]) <: Body

Construct a body by simply passing in the x and y coordinate vectors. The last point will be automatically connected to the first point. The coordinate vectors are assumed to be expressed in the body-fixed coordinate system. The optional closuretype specifies whether the body is closed (ClosedBody) or open (OpenBody). If closed, then the first and last points are assumed joined in operations that require neighbor points.

RigidBodyTools.Polygon — Type

Polygon(x::Vector,y::Vector,n[,closuretype=ClosedBody])

Create a polygon shape with vertices x and y, with approximately n points distributed along the perimeter.

RigidBodyTools.Circle — Function

Circle(a,n) <: Body

Construct a circular body with radius a and with n points distributed on the body perimeter.

Circle(a,targetsize::Float64) <: Body

Construct a circular body with radius a with spacing between points set approximately to targetsize.

RigidBodyTools.Ellipse — Type

Ellipse(a,b,n[,endpointson=false],[shifted=false]) <: Body

Construct an elliptical body with semi-major axis a and semi-minor axis b, with n points distributed on the body perimeter. Can also call this with Ellipse(a,b,ds) where ds is the desired spacing between points. If endpointson=true, then the endpoints will be placed on the surface and the midpoints will lie inside the nominal ellipse. By default, the midpoints lie on the ellipse. If shifted=true, then the first endpoint lies on an axis of symmetry. By default the first midpoint lies on the axis symmetry.

RigidBodyTools.NACA4 — Type

NACA4(cam,pos,thick,np,[len=1.0]) <: Body{N}

Generates points in the shape of a NACA 4-digit airfoil of chord length 1. The relative camber is specified by cam, the position of maximum camber (as fraction of chord) by pos, and the relative thickness by thick. The parameter np specifies the number of points on the upper or lower surface. The optional parameter len specifies the chord length, which defaults to 1.0.

Example

julia> b = NACA4(0.0,0.0,0.12);

RigidBodyTools.Plate — Function

Plate(a,n) <: Body

Construct a flat plate of zero thickness with length a, divided into n equal segments.

Plate(a,ds) <: Body

Construct a flat plate of zero thickness with length a, with approximate spacing ds between points.

RigidBodyTools.Rectangle — Function

Rectangle(a,b,n) <: Body

Construct a rectangular body with x̃ side half-length a and ỹ side half-length b, with approximately n points distributed along the perimeter. The centroid of the rectangle is placed at the origin (so that the lower left corner is at (-a,-b)).

Points are not placed at the corners, but rather, are shifted by half a segment. This ensures that all normals are perpendicular to the sides.

Rectangle(a,b,ds) <: Body

Construct a rectangular body with x̃ side half-length a and ỹ side half-length b, with approximate spacing ds between points.

RigidBodyTools.SplinedBody — Function

SplinedBody(X,Δx[,closuretype=ClosedBody]) -> BasicBody

Using control points in X (assumed to be N x 2, where N is the number of points), create a set of points that are uniformly spaced (with spacing Δx) on a curve that passes through the control points. A cubic parametric spline algorithm is used. If the optional parameter closuretype is set to OpenBody, then the end points are not joined together.

SplinedBody(x,y,Δx[,closuretype=ClosedBody]) -> BasicBody

Using control points in x and y, create a set of points that are uniformly spaced (with spacing Δx) on a curve that passes through the control points. A cubic parametric spline algorithm is used. If the optional parameter closuretype is set to OpenBody, then the end points are not joined together.

RigidBodyTools.Square — Function

Square(a,n) <: Body

Construct a square body with side half-length a and with approximately n points distributed along the perimeter.

Square(a,ds) <: Body

Construct a square body with side half-length a, with approximate spacing ds between points.

Rigid transformations of shapes

RigidBodyTools.RigidTransform — Type

RigidTransform(x::Tuple{Real,Real},α::Real)

Construct a rigid-body transform operator, with rotation by angle α and translation specified by x. The translation coordinates are specified in the target coordinate system.

The resulting transform can be used as an operator on pairs of coordinate vectors, x and y, or on bodies. For transformation of bodies, it only overwrites the x and y fields of the body, but leaves the x̃ and ỹ (body coordinates) intact.

The translation can be provided as either a tuple (x,y) or as a complex number.

Constructors

RigidTransform((x,y),α)

RigidTransform(u::Vector{Real})

RigidTransform(u::NTuple{3,Real})

RigidTransform.(u) where u is a collection of vectors or tuples.

Example

julia> body = RigidTransform.Ellipse(0.5,0.1,100)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

julia> T = RigidTransform((1.0,1.0),π/4)
Rigid-body transform
  Translation: (1.0,1.0)
  Rotation angle (rad): 0.7853981633974483

julia> T(body)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (1.0,1.0)
   Current angle (rad): 0.7853981633974483

Base.vec — Method

vec(T::RigidTransform) -> Vector{Float64}

Returns a length-3 vector of the form [x,y,α] corresponding to the translation and rotation specified by the given transform T.

Shape utilities

RigidBodyTools.centraldiff — Function

centraldiff(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the circular central differences of coordinates on body body (or on each body in list body). If ref=true, uses the reference coordinates in body-fixed space.

centraldiff(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the centraldiff on each constituent body in bl. If ref=true, uses the reference coordinates in body-fixed space.

Base.diff — Method

diff(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y differences of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body (if ref=false), or at the reference x̃ and ỹ coordinates (body-fixed space) if ref=true. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

Base.diff — Method

diff(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the diff on each constituent body in bl.

RigidBodyTools.dlength — Function

dlength(body::Body/BodyList[,ref=false]) -> Vector{Float64}

Compute the lengths of the faces on the perimeter of body body, whose ends are at the current xend and yend coordinates (in inertial space) of the body. Face 1 corresponds to the face between endpoints 1 and 2, for example. If ref=true, uses the reference coordinates in body-fixed space.

RigidBodyTools.dlengthmid — Function

dlengthmid(body::Body/BodyList[,ref=false]) -> Vector{Float64}

Same as dlength.

Base.length — Method

length(body::Body)

Return the number of points on the body perimeter

RigidBodyTools.midpoints — Method

midpoints(body::Body[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y midpoints of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body (if ref=false), or at the reference x̃ and ỹ coordinates (body-fixed space) if ref=true. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

RigidBodyTools.midpoints — Method

midpoints(bl::BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the midpoints on each constituent body in bl.

RigidBodyTools.normal — Function

normal(body::Body/BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals in inertial components (if ref=false) or body- fixed components (if ref=true) of the faces on the perimeter of body body, whose ends are at the current xend and yend coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example. For an OpenBody, this provides a vector that is one element shorter than the number of points.

RigidBodyTools.normalmid — Function

normalmid(body::Body/BodyList[,ref=false]) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals in inertial components (if ref=false) or body- fixed components (if ref=true) of the faces formed between endpoints on the perimeter of body body (or each body in list body).

RigidBodyTools.arccoord — Function

arccoord(body::Body/BodyList[,ref=false]) -> Vector{Float64}

Returns a vector containing the arclength coordinate along the surface of body, evaluated at the second endpoint of each face. So, e.g., the first coordinate would be the length of face 1, the second the length of face 2, and the last would be total length of all of the faces. Use inertial components (if ref=false) or body-fixed components (if ref=true). If this is a body list, restart the origin of the coordinates on each body in the list.

RigidBodyTools.arccoordmid — Function

arccoordmid(body::Body/BodyList[,ref=false]) -> Vector{Float64}

Returns a vector containing the arclength coordinate along the surface of body, evaluated at the midpoints between the ends of faces. So, e.g., the first coordinate would be half of the length of face 1, the second would be half of face 2 plus all of face 1, etc. Use inertial components (if ref=false) or body- fixed components (if ref=true). If this is a body list, restart the origin of the coordinates on each body in the list.

RigidBodyTools.arclength — Method

arclength(body::Body[,ref=false])

Compute the total arclength of body, from the sum of the lengths of the faces. If ref=true, use the body-fixed coordinates.

RigidBodyTools.arclength — Method

arclength(bl::BodyList[,ref=false]) -> Vector{Float64}

Compute the total arclength of each body in bl and assemble the results into a vector. If ref=true, use the body-fixed coordinates.

This page was generated using Literate.jl.